Integrated a feature request from a client.
The client uses the custom fields to track the location of assets
instead of checking out and moving them. They wanted a way to scan
a bunch of barcodes and then do a "mass edit" on them. I created
a "Aggregate Mass Edit" button on the asset list page that sends
you to a modified asset_edit page that allows the style of entry
that they wanted. Once the barcodes have been scanned it sends
you back to the asset_list page with the assets to edit pre-loaded
and opens up the mass edit window. When complete it sends you back
to the aggregate mass edit page.

Why didn't I make another button on the left like "Check out Assets"
or "Reserve Assets"? I tried this first. I had trouble opening the
"Mass Edit" dialog box when doing this and QCodo was crashing when
it tried to generate errors which made debugging difficult.
I decided on this other method of bouncing you around between pages
instead.